qtvcp -add checkboxes to tooloffsetview / code to delete tools
check the boxes and (in the dialog version) press delete and the checked tools
will be removed from the toolfile.

A ton of work to add space in the array for the checkbox.
pyqt is not nearly as easy as GTK in this regard.
This code needs clean up ... but that will have to come later