qtvcp -have axis selection widgets select the joint number too
When selecting a axis to un/home we nee the joint number set.
Since we are not in joint mode it was not set by the widget.
If we set it normally it sends a STATUS message that puts tings out of synch.
We get around this by setting the joint number without sending a STATUS message.
Hopefully that doesn't create problems later