qtvcp -More work on embedding programs
In the pursuit of having AXIS styles embedding.
QT5 doesn't do embedding well - very inconsistant.
Having qt embed a running program seems to work best.
Try that if we can, otherwise allow the program to embed.
There seems to be a race - hence the sleep statements

the move/resize event works only when qt embeds the program.
And it's flakey