Axis preview failing with wrapped rotary axis.
When a wrapped rotary axis absolute position is > 360 or < -360, axis preview will fail to load a file. This is because axis issues hidden "g53 g0 ..."  commands to the preview interpreter prior to sending the file contents to the preview interpreter. Axis uses the absolute coordinates as part of those commands which is not correct for a wrapped_rotary type axis.

Included is a sim file wrapped_rotary to demonstrate this issue as well as demonstrate the fix.

To Test:
  1. linuxcnc wrapped_rotary.ini
  2. home
  3. mdi "g0 a180"
  4. mdi "g0 a0"
  5. ctrl-r (reload axis.ngc)