qtvcp -add progressbar to overlay when loading gcode
In big programs it's nice to know it's still loading.
now overlay will hide the overlay instead of the dialog.
This is because dialog actually hides the overlay way before the program
is loaded, but because of python single thread it actually hides later.
But when we force an redraw in verlay to see the prgressbar change, it
changes this process.