emctaskmain.cc:  Fix stepping over first line
When running a program, the `programStartLine` is normally set to 0.

When the "run from line" option is set, then `programStartLine` is set
to something greater than zero.  Logic in `readahead_reading()` checks
if the currently-read line number is less or equal to than the "run
from" line, and if so, clears the interp queue before queueing up the
current line.

For unknown reasons dating back to 2005 (see c06400d4), when stepping,
the "run from" line is set to 1.  Normally, clearing the queue at this
point doesn't cause a problem, since nothing is queued up.  But in the
special case of a loop, any queued commands from previous iterations
are thrown away, and only the final iteration of the loop will
execute.

It seems this bug was only recently discovered by @andypugh after all
this time because of its obsucer, corner-case nature.