Fix `halcompile` and `uspace` tests against system install
These tests refer to files outside the `tests` directory.  To make the
tests function disembodied from the source tree and against a system
install, add rules to the build to copy the needed files in place, and
adjust the tests to work properly.

Also, the `personalities_mod` test did something funky:  it overwrote
standard comps with modified versions, then replaced those when the
test finished.  This patch undoes that, and instead renames the comps.