add interp pins containing extended state tag data
Adds a number of new tags to state_tag.h and publishes all state tags as pins
 to motion under the interp. section.

This commit also creates the interp.heading pin which depends on motion to
provide the current position. This could be used to contorl a tasngential
knife.

Many of these new added pins are useful for controlling plasma machines which
have unique requirements.