packaging: Do not require 'six' for doc building
Silent failures were occurring during the doc build because the
program `image-wildcard` could not be executed:
http://buildbot.linuxcnc.org/buildbot/builders/4042.deb-buster-rtpreempt-rpi4/builds/163/steps/shell_3/logs/stdio

    Traceback (most recent call last):
      File "../docs/src/image-wildcard", line 6, in <module>
        from six.moves import map
    ImportError: No module named six.moves

Rather than depending on six (which we _could_ do), just remove the
need for it.

While we're here, unconditionally use `relpath` just so we know
whether it works.

Testing performed: built html docs locally.  Inspected make output
and viewed axis.html; images were present.

Note:  There is no way I could find to make failures in asciidoc
`sys:` to cause asciidoc to indicate build failure back to make.
The only way to see such problems is apparently to read through
all the make output, or inspect the generated documentation files.

Note 2: This is likely to fail on buster-ish systems when imagemagic
is configured to block all reading and writing of PDF files, which was
recently done due to a known security problem in ghotscript:

    https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=964090