qtvcp -have qtvcp look for, compile and use resource files.
if there is a BASENAME.qrc file, it will (if required) compile it into a
resources.py file and then import it into qtvcp

it check to see if the qrc file is newer then the resources.py file.
If there is just a resources.py file, it will use it.
if there is neither, it will do lothing extra.