pncconf -fix setting of maxerror on metric machines
as per Peter Wallace:
This needs to be scaled correctly for metric.
Bounding the PID to 0.0005 mm doesn't give the PID enough headroom to adjust for
timebase differences between LinuxCNC and the stepgen hardware.

The point of this setting is to limit bogus stepgen velocity corrections caused by position
feedback sample time jitter. That is, say you have 200 usec of jitter on your servo thread and a
velocity of 10 IPS, a 200 usec late position sample would be off by 0.002", This would
cause the PID to command a bogus 20% velocity change with the standard PID tuning,
by bounding the maximum PID error to 0.0005" we get a 5% change instead.

Note that this is mostly obsolete if the DPLL is used as this reduces the sample time
jitter to about 250 ns.