qtvcp -simplify initializing programmed HALified widgets
We used to have to call hal_init() with a ton of variables.
Now they are embedded into an instance of _HalWidgetBase, so you just have
to call hal_init()
You can still use the variable HAL_NAME= to set the HAL pin name or
else it uses the pyqt objectName

update all the qtvcp code to use this new way.