uspace_rtapi_app: error-check use of pthread mutexes
It has been suggested that the logic in Posix::wait could lead to
an unlock of a not-locked mutex.  If this happens at runtime,
rtapi_app will now abort.

Note that as this lock is only used for sim / non-rt situations, this
cannot affect people running on hardware.

By setting RTAPI_LOCK_FAIL in the environment, you can check that the
checking works, by causing a "lock while already locked" type error.