gmoccapy -fix display of jogging buttons xyzab
The function to arange jogging buttons relies on the dict
that hold available axes to be in the same order as defined.
in python 2.7 that is not the default behaivor of dicts.
So the function (with a XYZAB machine for example) gave two
sets of row/coloumn numbers for four buttons, leading to two
buttons not being available.