qtvcp -fix HAL float input pins only getting s32
It seems on HAL float input pins that got their value
from QPin's QTSignal, would get an integer number rather then the float.
I could change that by changing the order of the overloaded pyqtSignal
definition but then that may have just made eveything a float...

Now we send a PyQt_PyObject that is not supposed to be converted
internally. It seems to work.