halcmd: fix "tune" command locking.
The "tune" command wasn't actually applying the desired locks
because it was locking a bitwise 'and' of values 1 & 2, which
gives a value of 0. This should have been a bitwise or of the
values.
Also changed the associated 'unlock' functions so that the "tune"
command unlocks the same locks as it locked.
Created a new #define to represent the locks to locks for the "tune"
command to help make it clearer that that the lock and unlock
functions should alter the same locks.