qtvcp -qt_istat: prebuild joint num to axis index conversions
dro_widget needs these conversions; num of joints vrs number of axes
are not nessasarily the same.
python linuxcnc library always reports axes positions in a set of 9 variables.
the number of joint number varies but is always the same or more then axes.

This makes a map of the differences.