qtvcp -dro_widget: fix a race condition with angular display
The widget checked for joint type early in the process by calling stat.
Stat might not have been polled yet so might get the wrong info.
This would show as the format of the angular DRO changing to imperial/metric
 format

It was noticed adding the touchyMDI widget would fix the problem
because it polled status it's self.