qtvcp -core: make HAL pin making crash proof
Now if pin making fails, it won;t kill the program
right away.
You can check for the error and an error is logged.
calls to set() are ignored.
probably will fail when get() is called.

Pobably sjould make it fail right away in verbose mode but
that will take changes in qtvcp.py