qyvcp -file_manader: add file details, scrolling via slots
now defults to a table of filename/timestamp and size.
Can be changed back to simple list

Can you designer signals to connect to the scroll slot.
This allows using a slider or diali widget to scroll up and down.

Coulfd also connect an external MPG to do the same thing.
QtTouchy sim shows how that would work (the code is a little different 0 - i've noe made it easier)