remap -add a tool-probe-at-m6 remap
This is a common modification to linuxcnc.
maybe if we have a 'standard' remap it will make things easier.
This is based on Versaprobe remap.
Param 4999 holds whether the machine in imperial.
param 5000 holds the workpiece height in absolute machine units.
The ini holds the rest of the required imforation.

To make things essier it would be nice to have a couple params
always available, such as what units the machine is based on
work in progress but it does work.

A toolsetter component for sim testing would be helpful too.

Better error handling of remaps, particular finding the files to import,
would really help people.
It it fails it would be nice to know where it looked...