qtvfcp -dro_widget: add a user display option.
Phill needed to display a DRO that was not linuxcnc defined.
He will drive it (externally) with a HAL pin.

you change the displayed valye by changing the qreference_type property
and calling .update_user() with a int/float value
You can also use that function as a slot for signals in designer