qtvcp -fix highlight selection on gcode display widgets
gcode graphics didn't send out the line that was selected.
gcode editor was broken when I fixed gcode graphics.

STATUS signals about gcode line are now consistant - lines are 1 based just
as the user displayed code is. Gcode editor is internally 0 based and this
caused a lot of problems.

Had to adjust the run-from-line toolbar for 1 based lines

added a user marker for toggling a visual point in gcode lines.
Now we use a marker for the gcode selected line rather then the carret
line. Finally!
can style selected text now (cause I was going to use that as highlight.