HAL: Increase the HAL memory size.
The current HAL memory limit is set at 85 * 4096 (340k) which allows around
2500 HAL pins to exist.
The limit was last increased in 2014 to accommodate longer HAL pin names.
Mesa smart-serial means that configurations are possible with many thousands
of pins, and  340k seems like a very small limit.
This patch increases the HAL memory to a full 1MB.


Signed-off-by: andypugh <andy@bodgesoc.org>