qtvcp -qtdragon/hd: make overriding limits less confusing
The override limits check button would not always be in synch with
linuxcnc. linuxcnc does not have a way to specicifically set the
override state, only toggle it.

between this and the machine on button color not following linuxcnc state,
could make it very confusing what was on or off.

Also linuxcnc override_limits seems to work slightly different in homed
vrs not homed state. In homed state the override stays active after a jog,
in not homed state, linuxcnc automatically sets the machine off and sets
the override off.

In qtvcp, when the hard limits are cleared, we automatically set
limits override to off.