qtvcp -fix saving of resource file
The resource file is compiled from the qrc file on demand.
In a installed version of linuxcnc, one can;t save the
compiled file back where the default qrc file is, without
being root.
Now we add a qtvcp directory to the users's config folder
and put the compiled resource file there.