rs274: Temporarily force the LC_NUMERIC setting to C as needed
Code in the gcode interpreter assumes that number parsing/formatting
is done according to the C locale, as gcode part programs always use
"." as the decimal separator.

However, some of the user interfaces set the LC_NUMERIC to the user's
locale and then call into the interpreter to prepare previews of the
part program.

The last time I addressed this, it was by changing one routine to work via
an ostringstream, which is not affected by locales by default.
This time I took a different approach, which is to temporarily force the
LC_NUMERIC to C at entry to certain functions, and to restore it on exit.

If UIs or other embedders of the interpreter also use threads, they
must use `uselocale` to switch from program-wide to per-thread locale
settings, or the behavior of their program becomes undefined.

This hopefully fixes multiple bugs in affected locales, and closes #1158.
However, I did not actually set up a testing environment that would allow
me to check it.  Additionally, saving & restoring parameters will now
take place in the C locale.