Increasing Hal Size as we Maxed out pin count.
On this ever increasing complexity of a build, Hal in formed us we were out of memory. `HAL: ERROR: Insufficient memory for signal 'j2-master-ctrl8'`. We increased this on our current repo. The Hal File that caused this has 330 nets, in addition to the built in Hal pins. It seems to be passing tests.