Revert "fix changelog date"
The timestamp was updated so the GitHub action tests passed. lintian
throws an error when the release date for the previous release, is newer
then the current release date.

This is likely to break every time the 2.8 branch gets updated and the
changelog is merged into master branch. Instead, we will update the
timestamp when we run GitHub action.

This reverts commit 7c4181bdba50fa4989528b2220ebbb3f00a52fa3.