qtvcp -dialog_widget: fix error with run-from-line openning on toolchange
We used a hack to open the details message on the tool change dialog.
I think on different versions of qt it clicked different buttons.
one user said the run-from-line dialog would show all by it's self.
this would be consistant with the assummed problem.
anyways we check a little harder to see what button we are pressing
with our hack