touchy -have Touchy conform more to postgui standards
Touchy has been able to use INI based postgui file for ages.
But it used to require touchy.hal was in the config folder even if it was empty.

Now it will still look for the touchy.hal file but won't error if it's not there.
We promote in the docs to use INI based postgui files.
IIRC Touchy is the only modern GUI that did postgui loading differently.
Now (postgui) instructions for all GUIs can be the same.