Fix "object has no attribute 'tolist'"
This occurs since the switch to pyopengl (yay!) but only if python3-numpy
is not also installed.

I was able to eliminate the conversion of the matrix to its 1D / list form
entirely.

I think I got pretty good coverage of everything _except_ qt vismach
in my testing, both with and without python3-numpy installed.  qt vismach
should be the same as tk vismach, as I made edits with the same intended
effect.