qyvcp -overlay: switch overlay to a window over the main window
originally we made the main window a parent of the overlay window.
Now the window is separate and moves/resizes with the main window.
unfortunately the window manager will stop the the overlay at the far edges.
I'n expecting that must people don't move the window when focus is on.