Use correct language code zh_CN for the Chinese documentation
The 'cn' code is for the country (China), while the language code is 'zh'.
This bring the language code in line with the code used in src/po/, which
already include zh_CN.po, zh_HK.po and zh_TW.po.

The linuxcnc-doc-cn package is renamed to linuxcnc-doc-zh-cn, as _ is not
valid in package names.  The new package is set to conflict and replace
on the old package.