halscope: Change logfile to csv format
The csv format has this specification:
    Comments is a hash '#'.
    Delimiter is set to semicolon ';'.
    Numbers is written with LC_NUMERICAL 'C', (1.2345).

In it's current state, the csv files does not conform with RFC 4180.

For now, the first line contains a comment, the second line includes the
headers, which will be pin and parameter names, followed by lines with values.

LC_NUMERIC is set to 'C' so all numbers will have the same format,
regardless of previous set locale.

It is possible to either keep the system installed locale for LC_NUMERICAL
or change the delimiter to comma `,`, not both.