halscope: Remove old and unused code regarding logging.
Most of this code was added in 2006-2007, but it was never finished.
I'm also not sure what the actual purpose was or the intention of all
the functionality.

It also cleans up the code in function `write_log_file` regarding
indentation and spacing.

Most of the now removed code was added with these commits:
commit 84eddc67f6360680f23595ebec6ac55a31228156
commit 44998d04200b282ee8c7fa3e91d6490f8fd4d8a7
commit 773ab03f98af7337b33ab1f9aa839937f9274348
commit 6ad70711305e791412edfc46d615320a974dce14