#1528 Short arcs of very large radius were being interpreted as full circles
The problem exists in both the motion planner and the graphical preview, and
each uses different logic to decide when the endpoints are close enough to
assume that a full circle was desired.

Given that circle start and end points are defined by cartesian coordinates
it makes little sense to use the arc internal angle to decide if a full
circle was intended. This patch determines if a full circle is intended based
only on the linear distance from start to end coordinates.
The logic behaves consistently (Preview and actual motion always now match)
but how close the start and end point need to be for the arc to be interpreted
as a full circle depends on the configuration base units.
(It will always be CART_FUZZ base units, currently 1.0e-8)

Signed-off-by: andypugh <andy@bodgesoc.org>