motmod refactoring: Remove joint and axis internal data from shared mem
This patch takes a first step in reducing the scope of internal data
of motmod by making joints' and axes' internal data inaccessible from
"the rest of the world".

Rationale: Reducing the scope of data makes the code easier to maintain
and extend. (Specifically, this is part of preparation for adding jerk
limited jog planning in the future.)