classicladder: Sort out problem with parameters in saved files.
It seems that several parameters related to Modbus has different
names when I compare our version with upstream, and the general Modbus
settings is saved in a different section. Reverted some of those changes
so we are backward compatible with our own files.

Noticed this when I created a .clp file with the old version, and opened
it in the new.