classicladder: Remove option for selecting '%I' to read to, with Modbus.
The '%I' and '%IW' variables are the physical input pins from HAL. They
will be overwritten by HAL, if they are ever written to by ClassicLadder.

The option is only removed from the GUI, which means that it is still
possible to make changes to the '.clp' file and still be able to write
to the variables.

In the earlier version, there was a comment in 'protocol_modbus_master.c'
in SetVarFromModbus(). This implies that previously, this was handled on
a lower level.