classicladder: Make config window narrower.
For screens with low resolution, the width of the config window could be
wider then the width of the screen. This commit decreases the width for
some of the labels, which in turn makes the window narrower.

There is a issue with labels and boxes, expanding, and taking up more
space then they should. Rewriting this to a grid might work, and at the
same time remove some deprecation warnings.