docs build: make po4a a little more verbose, so buildbot doesn't give up on it
It takes po4a around an hour to make documentation.pot on the slow
buildbot workers, and normally it does this job without producing any
output.  This makes the buildbot time out and assume the job is hung,
and it kills it and fails the build.

This commit adds -v to po4a's and msgmerge's command lines, which causes
them to emit a bit of progress information along the way.  This is enough
to convince the buildbot that those processes are not hung.