interpeter: allow full paths with o words and remaps
You can use the full path including having '~' expanded to the user path.

This should be useful with GUI code writing, we can search paths easier,
and just specify the path directly when calling the Oword.

Might also help new users to get a fully functioning cample config.
When switching from RIP to installed the paths usually get jumbled.
This means we could really use a link in USER/linuxcnc/ to all the sample code.

One remaining problem is if you give an incorrect path in MDI, it is not caught and
hangs milltask. I'm not sure of the MDI command process yet.