Correct parsing of HTML and Python files
The po4a.cfg rules to parse the HTML and Python files were set to
parse the files as manual pages by mistake.  Corrected this in po4a.cfg,
as well as converted the parsed HTML files to XHTML for more reliable
parsing, as the po4a parser is targeted at XHTML.

This make it easier to translate the HTML and Python files, and
require the POT file to be regenerated.