New component at_pidv2 based on at_pid and pid components.
Started with a copy of pid.c, and copied relevant code from
at_pid.c and inserted it into relevant locations.  This way
at_pidv2 become compatible with the current pid component, and
those that want to test autotuning can insert it in-place
without modifying any pins and signals in their HAL files.

The original at_pid code was commited by Peter G. Vavaroutsos,
so insert his name as author and also copy the module name over.

Manual page at_pidv2(9) is based on pid(9) with the comment about
the auto-tuning code not being updated since 2011 added and the
relevant parts coped over from at_pid(9).

This module is tested and found to be giving not completely unusable
P, I and D values for horizontal movements.