qtvcp -qt5_graphics: allow magic comments notify to print to terminal
(PREVIEW,notify, some test) in gcode should print the text.
Most of the code was present but at the time I didn't inderstand
what the code was used for. Now it will print to the terminal.
It could be overriden in gcode_graphics.py to do something else,
thought I'm not sure what it is really useful for - debugging?
it only prints when the gcode is loaded.