Only change spindle speed info and flush queue when speed changes
At the moment repeated G1 with identical S# setting will cause the
spindle movement to come to a halt between each G1.  This only make
sense when the spindle speed change, so add a test to avoid flushing
the work queue if the spindle speed do not change.

Also fixes incorrect indentation in the function.

Fixes #1813.