Have autogen.sh copy in some files
Debian policy calls for these files not to be duplicated. However, we
need to do it in a portable way (e.g., works on RPM-based systems so
/usr/share/misc/config.guess is not OK; does not touch network, so doing
curl/wget of the canonical config.guess file is not OK)

While there should be an obvious way to copy these files, there's .. not.
This is the best I could come up with, using files from automake without
actually invoking it.

While I did not actually BUILD on an rpm-based system, I checked the
content of the automake rpm of fedora 36 and it looks like it should work.

I'd rather avoid the new build-time dependency on automake, but
"automake --print-libdir" was the only 'portable' way I found to print
the location of a directory containing config.guess, config.sub and
install-sh.

Alternative to #1625