Add new test demonstrating endless loop with incorrect G71 parameters
I found this in issue #1146. This example g-code program causes
the AXIS LinuxCNC GUI to get completely stuck while its memory use explode.
I decided to write a small test case to demonstrate the problem.
Discovered that the program cause the rs274 program to generate an
endless stream of repeated instructions, which I suspect is the cause of
the memory explosion.

Flaggex xfail as it is expected to fail until the bug is found and
fixed.