Merge auto tune support into pid HAL component
Instead of having a separate PID component with auto tuning support,
it is easier for users to have only one PID component where auto
tuning is a runtime option for those that want to use it.

Drop the at_pid HAL component and move its implementation and
documentation into the pid HAL component.  The code is marked
using #ifdef AUTO_TUNING to make it easy to identify in the code,
and trivial to compile out for those that want to.

There is some questions on how useful and efficient the implemented
auto tuning method is, and while I am sure it can be improved, it
did work for my machie and provided useful Pgain, Igain and Dgain
values for my X and Y axis.  As it is a runtime option, those
that do not want automatic tuning can simply ignore this feature.

Keep the at_pid(9) manual page to let earlier users know to use
the pid component instead.