qtvcp -qt_action: CALL_MDI_WAIT now reliably catches errors
before if there were multiple MDI commands and if one errorred, the next was often
run regardless.
This was do to having two linuxcnc.error channels competing for the data.
We now use STATUS.ERROR only.
There may be more code that need to use STATUS error in Qtvcp  yet...