LCNC 2.8 - Pncconf - adding 7i96s
I know that the stable version only serves to remove the bug.

I would like to make an exception here. Reason:
I'm not a fan of the LPT port. I find a used 7i96 card + a stable version of LCNC to be the easiest way to reliably use LCNC.

When I started with LinuxCNC and the 7i96 card, its support was also only in the master version. I know that changing a few strings in the repositories is a very simple matter for good Linux users. When I started with Linux, it was hell for me.

I try to keep the KISS rule. Keep It Simply Stupid.
This should be the way to go for complete Linux beginners:
Buy a 7i96s card => install a stable version of LCNC => use Pncconf configurator to set LCNC => MILLING!!!
Believe me, some beginners are happy to be able to count the number of pulses/units.

If there was no shortage of chips for 7i96 cards, I would not have asked for this exception.

This configuration must not be merged into the master (2.9) branch.

I posted this configuration here but there was no response.
https://forum.linuxcnc.org/27-driver-boards/46356-can-i-use-7i96s-with-lcnc-2-8?start=20