Add reset pins for halui Rapid Override, Feed Override and Spindle Overrides
halui currently has pins for increasing and decreasing the above overrides but
there isn't currently a universal way to reset them back to 1.0.

My use case is, I have 3 rotary encoders on my control panel each with a
momentary centre push switch.
I have the encoders setup to adjust the overrides with the centre push switch
connected to the new reset pins to quickly set the overrides back to 100%.

Prior to making this change I could increase and decrease the overrides from my
control panel, but I had to use the GUI to reset back to 100%.