Use exec instead of exit $? in tests
Before this change, the shell process will linger around to extract the
exit value, only to return it to the parent process, and after
this change the task is left to the kernel and one less
process is used during testing.  This is a very small optimisation
of the tests to gut down the number of lines and reduce the resources
spend during testing.