uspace+rtai: don't unload rtai modules
This causes a low-probability crash at runtime, due to an apparent
rtai bug.

Initially introduce this only in the path of the rarely-used
uspace+rtai configuration. However, ultimately, either we could
move towards uspace+rtai as the default configuration for rtai
kernels, or we could apply the same workaround for kmod+rtai
mode.

This also makes it necessary for Load to not error if a module
is already loaded; this can be checked very easily under the modest
assumption that /sys is mounted.