Don't build tpcomp, it's only useful out-of-tree
.. the manpage is still generated.

This is done simply to get rid of the diagnostic message when it's built
in-tree.

I considered the alternative of providing a TOPDIR value during build,
but as I would only be able to test it during a uspace build I decided
not to do so. I'd be happy to see it done as a substitute, though, as
otherwise the component is likely to bitrot.