carousel: Silence diagnostic without fixing bug
I believe there is a bug in this component adjacent to this diagnostic,
which I've previously reported as #1943. Silence the warning by giving
an explicit initialization to align_pin.

This will definitely cause the edge-type carousel to get stuck in the
state machine at the point where it waits for the alignment signal.

Before this change, one possible undefined behavior was that register
and/or the byte(s) of stack corresponding to this local variable would
vary unpredictably, leading to the component progressing through the
steps of the state machine anyway. However, this should not be relied
on.