Replace Debian autopkgtest setup
The original runtests script failed because the source code is not
configured when it is running, and scripts/runtests do not exist.

Instad, replace it with two tests, one testing the rs274 program, and
another testing the linuxcnc program.  The latter is a copy of the
tests/single-step/ self test with its files renamed to avoid any
name conflicts with the rs274 and future tests.