rtapi_app: add a cycle-counter interface
This adds `cycle_counter_get_freq()` and
`cycle_counter_get_count()`.  These functions provide access to the
hardware cycle-counter functionality (TSC on x86 and CNTVCT on aarch64).

These cycle counter interfaces are super fast on aarch64 (Raspberry
Pi 4B): 20-200 ns (compared to 50-15,000 ns for `clock_gettime()` on
this platform).

The results are not quite as good on amd64.