hal_sampler: truncate sample output file when opening
Before this, `hal_sampler FILE` would open the specified file with
O_CREAT (create file if it does not exist) but without O_TRUNC and without
O_APPEND, which means if the file already existed we would start writing
at the *beginning* of the file, overwriting and replacing the old contents
byte by byte (specifically: not line by line).

This means we didn't write enough data to overwrite all the old contents
of the file, then the last of the old contents would remain at the end
of the file, after the new data we wrote, which is just confusing.

This commit turns on the O_TRUNC flag, truncating the file to 0 bytes
when opening an existing file, discarding all the old data.