Fix boost python detection on arm and hppa in ax_boost_base.m4.
This should fix the failing build on Debian.

Also reintroduce a typo fix in the file header that was lost in a recent
upgrade of the file version.

The change has been emailed Thomas Porschberg to get it upstream.  Not sure
if it is going to work, as a comment on
<URL: http://randspringer.de/boost/index.html > make it sound like he want
to no longer maintain the file.