Added new tests/classicladder/estop setup.
This tests the estop behavour in classicladder, pressing an external
estop button, checking estop activates, and then reset estop and ensure
it clears.

Based on configs/sim/axis/classicladder/, with lube related setup and other
parts that seem to be irrelevant for this test removed.

The test remove use of eatmydata, as this fail for no apparent reason with
this error message:

  ERROR: ld.so: object 'libeatmydata.so' from LD_PRELOAD cannot be
         preloaded (cannot open shared object file): ignored.