hostmot2: collect initial writes into a single packet
This initial writes to the interface card happens with a series of small
writes that are all sent back-to-back. At least in the case of the gbit-
ethernet card litehm2 this can lead to packet loss, leading to a failed
startup.
This patch collects all packets generated by the force_write code into a
single packet.