task: handle pluggable interpreters better
The INI file may override which interpreter Task uses, via
[TASK]INTERPRETER.

Before this commit, if the user's specified interpreter failed to load,
Task would log a warning but then fall back to using the default interp
and continue.

After this commit, if the user's specified interpreter fails to load,
Task refuses to start.